adding re-run a check
